Added timer
Basically count down implemented with base unit being milliseconds

also introduced mux control to control which register is being displayed. It should be noted that a blinking was implemented so that both the active clock (real time or timer) would flash between which ever register was mbeing manually set, this looks awful and is confusing this needs to be changed.